+++ Sources/Moya/MoyaProvider+Defaults.swift
@@ -5,11 +5,12 @@ import Alamofire
 public extension MoyaProvider {
     public final class func defaultEndpointMapping(for target: Target) -> Endpoint<Target> {
         return Endpoint(
-            url: url(for: target).absoluteString,
+            url: URL(target: target).absoluteString,
             sampleResponseClosure: { .networkResponse(200, target.sampleData) },
             method: target.method,
             parameters: target.parameters,
-            parameterEncoding: target.parameterEncoding
+            parameterEncoding: target.parameterEncoding,
+            httpHeaderFields: target.headers
         )
     }
 
@@ -28,16 +29,5 @@ public extension MoyaProvider {
         let manager = Manager(configuration: configuration)
         manager.startRequestsImmediately = false
         return manager
-    }
-
-    // When a TargetType's path is empty, URL.appendingPathComponent may introduce trailing /, which may not be wanted in some cases
-    // See: https://github.com/Moya/Moya/pull/1053
-    // And: https://github.com/Moya/Moya/issues/1049
-    private final class func url(for target: Target) -> URL {
-        if target.path.isEmpty {
-            return target.baseURL
-        }
-
-        return target.baseURL.appendingPathComponent(target.path)
     }
 }