+++ Sources/RxMoya/MoyaProvider+Rx.swift
@@ -4,29 +4,31 @@ import RxSwift
 import Moya
 #endif
 
-/// Subclass of MoyaProvider that returns Observable instances when requests are made. Much better than using completion closures.
-open class RxMoyaProvider<Target>: MoyaProvider<Target> where Target: TargetType {
-    /// Initializes a reactive provider.
-    public override init(endpointClosure: @escaping EndpointClosure = MoyaProvider.defaultEndpointMapping,
-                         requestClosure: @escaping RequestClosure = MoyaProvider.defaultRequestMapping,
-                         stubClosure: @escaping StubClosure = MoyaProvider.neverStub,
-                         callbackQueue: DispatchQueue? = nil,
-                         manager: Manager = RxMoyaProvider<Target>.defaultAlamofireManager(),
-                         plugins: [PluginType] = [],
-                         trackInflights: Bool = false) {
-        super.init(endpointClosure: endpointClosure, requestClosure: requestClosure, stubClosure: stubClosure, callbackQueue: callbackQueue, manager: manager, plugins: plugins, trackInflights: trackInflights)
-    }
+extension MoyaProvider: ReactiveCompatible {}
+
+public extension Reactive where Base: MoyaProviderType {
 
     /// Designated request-making method.
     ///
     /// - Parameters:
     ///   - token: Entity, which provides specifications necessary for a `MoyaProvider`.
     ///   - callbackQueue: Callback queue. If nil - queue from provider initializer will be used.
-    /// - Returns: Cold observable, which emits one element or error.
-    open func request(_ token: Target, callbackQueue: DispatchQueue? = nil) -> Observable<Response> {
-        // Creates an observable that starts a request each time it's subscribed to.
-        return Observable.create { observer in
-            let cancellableToken = self.request(token, callbackQueue: callbackQueue) { result in
+    /// - Returns: Single response object.
+    public func request(_ token: Base.Target, callbackQueue: DispatchQueue? = nil) -> Single<Response> {
+        return base.rxRequest(token, callbackQueue: callbackQueue)
+    }
+
+    /// Designated request-making method with progress.
+    public func requestWithProgress(_ token: Base.Target, callbackQueue: DispatchQueue? = nil) -> Observable<ProgressResponse> {
+        return base.rxRequestWithProgress(token, callbackQueue: callbackQueue)
+    }
+}
+
+internal extension MoyaProviderType {
+
+    internal func rxRequest(_ token: Target, callbackQueue: DispatchQueue? = nil) -> Single<Response> {
+        return Observable.create { [weak self] observer in
+            let cancellableToken = self?.request(token, callbackQueue: callbackQueue, progress: nil) { result in
                 switch result {
                 case let .success(response):
                     observer.onNext(response)
@@ -37,14 +39,12 @@ open class RxMoyaProvider<Target>: MoyaProvider<Target> where Target: TargetType
             }
 
             return Disposables.create {
-                cancellableToken.cancel()
+                cancellableToken?.cancel()
             }
-        }
+        }.asSingle()
     }
-}
 
-public extension RxMoyaProvider {
-    public func requestWithProgress(_ token: Target, callbackQueue: DispatchQueue? = nil) -> Observable<ProgressResponse> {
+    internal func rxRequestWithProgress(_ token: Target, callbackQueue: DispatchQueue? = nil) -> Observable<ProgressResponse> {
         let progressBlock: (AnyObserver) -> (ProgressResponse) -> Void = { observer in
             return { progress in
                 observer.onNext(progress)