+++ Tests/NetworkLoggerPluginSpec.swift
@@ -21,7 +21,7 @@ final class NetworkLoggerPluginSpec: QuickSpec {
             let string: String = stringArray.reduce("") { $0 + $1 + " " }
             log += string
         })
-		
+
         let pluginWithRequestDataFormatter = NetworkLoggerPlugin(verbose: true, output: { (_, _, printing: Any...) in
             //mapping the Any... from items to a string that can be compared
             let stringArray: [String] = printing.map { $0 as? String }.flatMap { $0 }
@@ -72,7 +72,7 @@ final class NetworkLoggerPluginSpec: QuickSpec {
         }
 
         it("outputs the response data") {
-            let response = Response(statusCode: 200, data: "cool body".data(using: .utf8)!, response: URLResponse(url: URL(string: url(GitHub.zen))!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil))
+            let response = Response(statusCode: 200, data: "cool body".data(using: .utf8)!, response: HTTPURLResponse(url: URL(string: url(GitHub.zen))!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil))
             let result: Result<Moya.Response, MoyaError> = .success(response)
 
             plugin.didReceive(result, target: GitHub.zen)
@@ -83,7 +83,7 @@ final class NetworkLoggerPluginSpec: QuickSpec {
         }
 
         it("outputs the formatted response data") {
-            let response = Response(statusCode: 200, data: "cool body".data(using: .utf8)!, response: URLResponse(url: URL(string: url(GitHub.zen))!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil))
+            let response = Response(statusCode: 200, data: "cool body".data(using: .utf8)!, response: HTTPURLResponse(url: URL(string: url(GitHub.zen))!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil))
             let result: Result<Moya.Response, MoyaError> = .success(response)
 
             pluginWithResponseDataFormatter.didReceive(result, target: GitHub.zen)
@@ -92,13 +92,13 @@ final class NetworkLoggerPluginSpec: QuickSpec {
             expect(log).to( contain("{ URL: https://api.github.com/zen }") )
             expect(log).to( contain("formatted body") )
         }
-		
+
         it("outputs the formatted request data") {
-            let response = Response(statusCode: 200, data: "cool body".data(using: .utf8)!, response: URLResponse(url: URL(string: url(GitHub.zen))!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil))
+            let response = Response(statusCode: 200, data: "cool body".data(using: .utf8)!, response: HTTPURLResponse(url: URL(string: url(GitHub.zen))!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil))
             let result: Result<Moya.Response, MoyaError> = .success(response)
-            
+
             pluginWithRequestDataFormatter.didReceive(result, target: GitHub.zen)
-            
+
             expect(log).to( contain("Response:") )
             expect(log).to( contain("{ URL: https://api.github.com/zen }") )
             expect(log).to( contain("formatted request body") )
@@ -106,7 +106,7 @@ final class NetworkLoggerPluginSpec: QuickSpec {
 
         it("outputs an empty response message") {
             let emptyResponseError = AFError.responseSerializationFailed(reason: .inputDataNil)
-            let result: Result<Moya.Response, MoyaError> = .failure(.underlying(emptyResponseError))
+            let result: Result<Moya.Response, MoyaError> = .failure(.underlying(emptyResponseError, nil))
 
             plugin.didReceive(result, target: GitHub.zen)